Revert "chore: warehouse tests race detection (#3773)"
This reverts commit ce2ed332a45b3deafc7284e2629fb21a3f756aed.
