+++ warehouse/logfield/logfield.go
@@ -29,6 +29,5 @@ const (
 	QueryExecutionTime         = "queryExecutionTime"
 	StagingTableName           = "stagingTableName"
 	TotalRows                  = "totalRows"
-	WarehouseID                = "warehouseID"
 	SampleDuplicateMessages    = "sampleDuplicateMessages"
 )
