+++ warehouse/schema_test.go
@@ -1,4 +1,4 @@
-package schema
+package warehouse
 
 import (
 	"context"
@@ -6,14 +6,276 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/samber/lo"
-	"github.com/stretchr/testify/require"
-
 	"github.com/rudderlabs/rudder-go-kit/logger"
 	backendconfig "github.com/rudderlabs/rudder-server/backend-config"
+
+	"github.com/samber/lo"
+
 	"github.com/rudderlabs/rudder-server/warehouse/internal/model"
-	whutils "github.com/rudderlabs/rudder-server/warehouse/utils"
+
+	"github.com/stretchr/testify/require"
+
+	warehouseutils "github.com/rudderlabs/rudder-server/warehouse/utils"
 )
+
+func TestHandleSchemaChange(t *testing.T) {
+	inputs := []struct {
+		name             string
+		existingDatatype string
+		currentDataType  string
+		value            any
+
+		newColumnVal any
+		convError    error
+	}{
+		{
+			name:             "should send int values if existing datatype is int, new datatype is float",
+			existingDatatype: "int",
+			currentDataType:  "float",
+			value:            1.501,
+			newColumnVal:     1,
+		},
+		{
+			name:             "should send float values if existing datatype is float, new datatype is int",
+			existingDatatype: "float",
+			currentDataType:  "int",
+			value:            1,
+			newColumnVal:     1.0,
+		},
+		{
+			name:             "should send string values if existing datatype is string, new datatype is boolean",
+			existingDatatype: "string",
+			currentDataType:  "boolean",
+			value:            false,
+			newColumnVal:     "false",
+		},
+		{
+			name:             "should send string values if existing datatype is string, new datatype is int",
+			existingDatatype: "string",
+			currentDataType:  "int",
+			value:            1,
+			newColumnVal:     "1",
+		},
+		{
+			name:             "should send string values if existing datatype is string, new datatype is float",
+			existingDatatype: "string",
+			currentDataType:  "float",
+			value:            1.501,
+			newColumnVal:     "1.501",
+		},
+		{
+			name:             "should send string values if existing datatype is string, new datatype is datetime",
+			existingDatatype: "string",
+			currentDataType:  "datetime",
+			value:            "2022-05-05T00:00:00.000Z",
+			newColumnVal:     "2022-05-05T00:00:00.000Z",
+		},
+		{
+			name:             "should send string values if existing datatype is string, new datatype is string",
+			existingDatatype: "string",
+			currentDataType:  "json",
+			value:            `{"json":true}`,
+			newColumnVal:     `{"json":true}`,
+		},
+		{
+			name:             "should send json string values if existing datatype is json, new datatype is boolean",
+			existingDatatype: "json",
+			currentDataType:  "boolean",
+			value:            false,
+			newColumnVal:     "false",
+		},
+		{
+			name:             "should send json string values if existing datatype is jso, new datatype is int",
+			existingDatatype: "json",
+			currentDataType:  "int",
+			value:            1,
+			newColumnVal:     "1",
+		},
+		{
+			name:             "should send json string values if existing datatype is json, new datatype is float",
+			existingDatatype: "json",
+			currentDataType:  "float",
+			value:            1.501,
+			newColumnVal:     "1.501",
+		},
+		{
+			name:             "should send json string values if existing datatype is json, new datatype is json",
+			existingDatatype: "json",
+			currentDataType:  "datetime",
+			value:            "2022-05-05T00:00:00.000Z",
+			newColumnVal:     `"2022-05-05T00:00:00.000Z"`,
+		},
+		{
+			name:             "should send json string values if existing datatype is json, new datatype is string",
+			existingDatatype: "json",
+			currentDataType:  "string",
+			value:            "string value",
+			newColumnVal:     `"string value"`,
+		},
+		{
+			name:             "should send json string values if existing datatype is json, new datatype is array",
+			existingDatatype: "json",
+			currentDataType:  "array",
+			value:            []any{false, 1, "string value"},
+			newColumnVal:     []any{false, 1, "string value"},
+		},
+		{
+			name:             "existing datatype is boolean, new datatype is int",
+			existingDatatype: "boolean",
+			currentDataType:  "int",
+			value:            1,
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is boolean, new datatype is float",
+			existingDatatype: "boolean",
+			currentDataType:  "float",
+			value:            1.501,
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is boolean, new datatype is string",
+			existingDatatype: "boolean",
+			currentDataType:  "string",
+			value:            "string value",
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is boolean, new datatype is datetime",
+			existingDatatype: "boolean",
+			currentDataType:  "datetime",
+			value:            "2022-05-05T00:00:00.000Z",
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is boolean, new datatype is json",
+			existingDatatype: "boolean",
+			currentDataType:  "json",
+			value:            `{"json":true}`,
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is int, new datatype is boolean",
+			existingDatatype: "int",
+			currentDataType:  "boolean",
+			value:            false,
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is int, new datatype is string",
+			existingDatatype: "int",
+			currentDataType:  "string",
+			value:            "string value",
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is int, new datatype is datetime",
+			existingDatatype: "int",
+			currentDataType:  "datetime",
+			value:            "2022-05-05T00:00:00.000Z",
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is int, new datatype is json",
+			existingDatatype: "int",
+			currentDataType:  "json",
+			value:            `{"json":true}`,
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is int, new datatype is float",
+			existingDatatype: "int",
+			currentDataType:  "float",
+			value:            1,
+			convError:        errIncompatibleSchemaConversion,
+		},
+		{
+			name:             "existing datatype is float, new datatype is boolean",
+			existingDatatype: "float",
+			currentDataType:  "boolean",
+			value:            false,
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is float, new datatype is int",
+			existingDatatype: "float",
+			currentDataType:  "int",
+			value:            1.0,
+			convError:        errIncompatibleSchemaConversion,
+		},
+		{
+			name:             "existing datatype is float, new datatype is string",
+			existingDatatype: "float",
+			currentDataType:  "string",
+			value:            "string value",
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is float, new datatype is datetime",
+			existingDatatype: "float",
+			currentDataType:  "datetime",
+			value:            "2022-05-05T00:00:00.000Z",
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is float, new datatype is json",
+			existingDatatype: "float",
+			currentDataType:  "json",
+			value:            `{"json":true}`,
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is datetime, new datatype is boolean",
+			existingDatatype: "datetime",
+			currentDataType:  "boolean",
+			value:            false,
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is datetime, new datatype is string",
+			existingDatatype: "datetime",
+			currentDataType:  "string",
+			value:            "string value",
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is datetime, new datatype is int",
+			existingDatatype: "datetime",
+			currentDataType:  "int",
+			value:            1,
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is datetime, new datatype is float",
+			existingDatatype: "datetime",
+			currentDataType:  "float",
+			value:            1.501,
+			convError:        errSchemaConversionNotSupported,
+		},
+		{
+			name:             "existing datatype is datetime, new datatype is json",
+			existingDatatype: "datetime",
+			currentDataType:  "json",
+			value:            `{"json":true}`,
+			convError:        errSchemaConversionNotSupported,
+		},
+	}
+	for _, ip := range inputs {
+		tc := ip
+
+		t.Run(tc.name, func(t *testing.T) {
+			t.Parallel()
+
+			newColumnVal, convError := handleSchemaChange(
+				model.SchemaType(tc.existingDatatype),
+				model.SchemaType(tc.currentDataType),
+				tc.value,
+			)
+			require.Equal(t, newColumnVal, tc.newColumnVal)
+			require.ErrorIs(t, convError, tc.convError)
+		})
+	}
+}
 
 type mockSchemaRepo struct {
 	err       error
@@ -60,7 +322,7 @@ func TestSchema_GetUpdateLocalSchema(t *testing.T) {
 		sourceID      = "test_source"
 		destID        = "test_dest"
 		namespace     = "test_namespace"
-		warehouseType = whutils.RS
+		warehouseType = warehouseutils.RS
 		uploadID      = 1
 	)
 
@@ -139,15 +401,14 @@ func TestSchema_GetUpdateLocalSchema(t *testing.T) {
 
 			ctx := context.Background()
 
-			err := sch.UpdateLocalSchema(ctx, uploadID, tc.mockSchema.Schema)
+			err := sch.updateLocalSchema(ctx, uploadID, tc.mockSchema.Schema)
 			if tc.wantError == nil {
 				require.NoError(t, err)
 			} else {
 				require.ErrorContains(t, err, tc.wantError.Error())
 			}
 
-			localSchema, err := sch.GetLocalSchema(ctx)
-			require.Equal(t, tc.wantSchema, localSchema)
+			err = sch.fetchSchemaFromLocal(ctx)
 			require.Equal(t, tc.wantSchema, sch.localSchema)
 			if tc.wantError == nil {
 				require.NoError(t, err)
@@ -159,6 +420,8 @@ func TestSchema_GetUpdateLocalSchema(t *testing.T) {
 }
 
 func TestSchema_FetchSchemaFromWarehouse(t *testing.T) {
+	Init4()
+
 	testCases := []struct {
 		name           string
 		mockSchema     model.Schema
@@ -283,7 +546,7 @@ func TestSchema_FetchSchemaFromWarehouse(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 
-			fetchSchemaRepo := mockFetchSchemaFromWarehouse{
+			fechSchemaRepo := mockFetchSchemaFromWarehouse{
 				schemaInWarehouse:             tc.mockSchema,
 				unrecognizedSchemaInWarehouse: tc.mockSchema,
 				err:                           tc.mockErr,
@@ -308,15 +571,13 @@ func TestSchema_FetchSchemaFromWarehouse(t *testing.T) {
 
 			ctx := context.Background()
 
-			schemaInWarehouse, unrecognizedSchemaInWarehouse, err := sh.FetchSchemaFromWarehouse(ctx, &fetchSchemaRepo)
+			err := sh.fetchSchemaFromWarehouse(ctx, &fechSchemaRepo)
 			if tc.wantError != nil {
 				require.EqualError(t, err, tc.wantError.Error())
 			} else {
 				require.NoError(t, err)
 			}
-			require.Equal(t, tc.expectedSchema, schemaInWarehouse)
 			require.Equal(t, tc.expectedSchema, sh.schemaInWarehouse)
-			require.Equal(t, tc.expectedSchema, unrecognizedSchemaInWarehouse)
 			require.Equal(t, tc.expectedSchema, sh.unrecognizedSchemaInWarehouse)
 		})
 	}
@@ -324,17 +585,18 @@ func TestSchema_FetchSchemaFromWarehouse(t *testing.T) {
 
 func TestSchema_GetUploadSchemaDiff(t *testing.T) {
 	testCases := []struct {
-		name              string
-		tableName         string
-		currentSchema     model.Schema
-		uploadTableSchema model.TableSchema
-		expected          whutils.TableSchemaDiff
+		name          string
+		tableName     string
+		currentSchema model.Schema
+		uploadSchema  model.Schema
+		expected      warehouseutils.TableSchemaDiff
 	}{
 		{
-			name:          "empty current and upload table schema",
+			name:          "empty current and upload schema",
 			tableName:     "test-table",
 			currentSchema: model.Schema{},
-			expected: whutils.TableSchemaDiff{
+			uploadSchema:  model.Schema{},
+			expected: warehouseutils.TableSchemaDiff{
 				ColumnMap:        model.TableSchema{},
 				UpdatedSchema:    model.TableSchema{},
 				AlteredColumnMap: model.TableSchema{},
@@ -344,10 +606,12 @@ func TestSchema_GetUploadSchemaDiff(t *testing.T) {
 			name:          "empty current schema",
 			tableName:     "test-table",
 			currentSchema: model.Schema{},
-			uploadTableSchema: model.TableSchema{
-				"test-column": "test-value",
+			uploadSchema: model.Schema{
+				"test-table": model.TableSchema{
+					"test-column": "test-value",
+				},
 			},
-			expected: whutils.TableSchemaDiff{
+			expected: warehouseutils.TableSchemaDiff{
 				Exists:           true,
 				TableToBeCreated: true,
 				ColumnMap: model.TableSchema{
@@ -367,10 +631,12 @@ func TestSchema_GetUploadSchemaDiff(t *testing.T) {
 					"test-column": "test-value-1",
 				},
 			},
-			uploadTableSchema: model.TableSchema{
-				"test-column": "test-value-2",
+			uploadSchema: model.Schema{
+				"test-table": model.TableSchema{
+					"test-column": "test-value-2",
+				},
 			},
-			expected: whutils.TableSchemaDiff{
+			expected: warehouseutils.TableSchemaDiff{
 				Exists:           false,
 				TableToBeCreated: false,
 				ColumnMap:        model.TableSchema{},
@@ -389,10 +655,12 @@ func TestSchema_GetUploadSchemaDiff(t *testing.T) {
 					"test-column-2": "test-value-2",
 				},
 			},
-			uploadTableSchema: model.TableSchema{
-				"test-column": "test-value-2",
+			uploadSchema: model.Schema{
+				"test-table": model.TableSchema{
+					"test-column": "test-value-2",
+				},
 			},
-			expected: whutils.TableSchemaDiff{
+			expected: warehouseutils.TableSchemaDiff{
 				Exists:           true,
 				TableToBeCreated: false,
 				ColumnMap: model.TableSchema{
@@ -415,10 +683,12 @@ func TestSchema_GetUploadSchemaDiff(t *testing.T) {
 					"test-column-2": "test-value-2",
 				},
 			},
-			uploadTableSchema: model.TableSchema{
-				"test-column": "text",
+			uploadSchema: model.Schema{
+				"test-table": model.TableSchema{
+					"test-column": "text",
+				},
 			},
-			expected: whutils.TableSchemaDiff{
+			expected: warehouseutils.TableSchemaDiff{
 				Exists:           true,
 				TableToBeCreated: false,
 				ColumnMap:        model.TableSchema{},
@@ -440,8 +710,9 @@ func TestSchema_GetUploadSchemaDiff(t *testing.T) {
 
 			sch := Schema{
 				schemaInWarehouse: tc.currentSchema,
+				uploadSchema:      tc.uploadSchema,
 			}
-			diff := sch.TableSchemaDiff(tc.tableName, tc.uploadTableSchema)
+			diff := sch.generateTableSchemaDiff(tc.tableName)
 			require.EqualValues(t, diff, tc.expected)
 		})
 	}
@@ -593,17 +864,19 @@ func TestSchema_HasLocalSchemaChanged(t *testing.T) {
 
 			sch := &Schema{
 				warehouse: model.Warehouse{
-					Type: whutils.SNOWFLAKE,
+					Type: warehouseutils.SNOWFLAKE,
 				},
 				skipDeepEqualSchemas: tc.skipDeepEquals,
+				localSchema:          tc.localSchema,
+				schemaInWarehouse:    tc.schemaInWarehouse,
 			}
-			require.Equal(t, tc.expected, sch.hasSchemaChanged(tc.localSchema, tc.schemaInWarehouse))
+			require.Equal(t, tc.expected, sch.hasSchemaChanged())
 		})
 	}
 }
 
 func TestSchema_PrepareUploadSchema(t *testing.T) {
-	whutils.Init()
+	warehouseutils.Init()
 
 	const (
 		sourceID    = "test-source-id"
@@ -633,7 +906,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 	}{
 		{
 			name:          "error fetching staging schema",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas:   []model.Schema{},
 			mockErr:       errors.New("test error"),
 			wantError:     errors.New("consolidating staging files schema: getting staging files schema: test error"),
@@ -641,7 +914,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 
 		{
 			name:          "discards schema for bigquery",
-			warehouseType: whutils.BQ,
+			warehouseType: warehouseutils.BQ,
 			mockSchemas:   []model.Schema{},
 			expectedSchema: model.Schema{
 				"rudder_discards": model.TableSchema{
@@ -657,7 +930,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "discards schema for all destinations",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas:   []model.Schema{},
 			expectedSchema: model.Schema{
 				"rudder_discards": model.TableSchema{
@@ -672,7 +945,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "users and identifies should have similar schema except for id and user_id",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"identifies": model.TableSchema{
@@ -736,7 +1009,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "users have extra properties as compared to identifies",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"identifies": model.TableSchema{
@@ -804,7 +1077,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "users without identifies",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"users": model.TableSchema{
@@ -844,7 +1117,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "unknown table in warehouse schema",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"test-table": model.TableSchema{
@@ -888,7 +1161,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "unknown properties in warehouse schema",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"test-table": model.TableSchema{
@@ -932,7 +1205,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "single staging schema with empty warehouse schema",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"test-table": model.TableSchema{
@@ -966,7 +1239,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "single staging schema with warehouse schema and text data type override",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"test-table": model.TableSchema{
@@ -1013,7 +1286,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:                "id resolution without merge schema",
-			warehouseType:       whutils.BQ,
+			warehouseType:       warehouseutils.BQ,
 			idResolutionEnabled: true,
 			mockSchemas: []model.Schema{
 				{
@@ -1049,7 +1322,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:                "id resolution with merge schema",
-			warehouseType:       whutils.BQ,
+			warehouseType:       warehouseutils.BQ,
 			idResolutionEnabled: true,
 			mockSchemas: []model.Schema{
 				{
@@ -1099,7 +1372,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "multiple staging schemas with empty warehouse schema",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"test-table-1": model.TableSchema{
@@ -1151,7 +1424,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "multiple staging schemas with empty warehouse schema and text datatype",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"test-table-1": model.TableSchema{
@@ -1227,7 +1500,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "multiple staging schemas with warehouse schema and text datatype",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"test-table-1": model.TableSchema{
@@ -1315,7 +1588,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "multiple schemas with same table and empty warehouse schema",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"test-table": model.TableSchema{
@@ -1353,7 +1626,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "multiple schemas with same table and warehouse schema",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"test-table": model.TableSchema{
@@ -1397,7 +1670,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "multiple schemas with preference to first schema and empty warehouse schema",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"test-table": model.TableSchema{
@@ -1441,7 +1714,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "multiple schemas with preference to warehouse schema",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"test-table": model.TableSchema{
@@ -1495,7 +1768,7 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 		},
 		{
 			name:          "warehouse users have extra properties as compared to identifies",
-			warehouseType: whutils.RS,
+			warehouseType: warehouseutils.RS,
 			mockSchemas: []model.Schema{
 				{
 					"identifies": model.TableSchema{
@@ -1628,13 +1901,13 @@ func TestSchema_PrepareUploadSchema(t *testing.T) {
 				stagingFilesSchemaPaginationSize: 2,
 			}
 
-			uploadSchema, err := sh.ConsolidateLocalSchemaWithStagingFiles(ctx, stagingFiles)
+			err := sh.prepareUploadSchema(ctx, stagingFiles)
 			if tc.wantError != nil {
 				require.EqualError(t, err, tc.wantError.Error())
 			} else {
 				require.NoError(t, err)
 			}
-			require.Equal(t, tc.expectedSchema, uploadSchema)
+			require.Equal(t, tc.expectedSchema, sh.uploadSchema)
 		})
 	}
 }
