+++ warehouse/upload_test.go
@@ -9,7 +9,6 @@ import (
 	"time"
 
 	"github.com/rudderlabs/rudder-go-kit/stats"
-	"github.com/rudderlabs/rudder-server/warehouse/schema"
 
 	"github.com/ory/dockertest/v3"
 
@@ -153,13 +152,16 @@ func TestColumnCountStat(t *testing.T) {
 					},
 				},
 				statsFactory: store,
-				schemaHandle: &schema.Schema{},
+				schemaHandle: &Schema{
+					schemaInWarehouse: model.Schema{
+						tableName: model.TableSchema{
+							"test-column-1": "string",
+							"test-column-2": "string",
+							"test-column-3": "string",
+						},
+					},
+				},
 			}
-			j.schemaHandle.SetWarehouseTableSchema(tableName, model.TableSchema{
-				"test-column-1": "string",
-				"test-column-2": "string",
-				"test-column-3": "string",
-			})
 
 			tags := j.buildTags()
 			tags["tableName"] = tableName
@@ -170,7 +172,7 @@ func TestColumnCountStat(t *testing.T) {
 			m2 := store.Get("warehouse_load_table_column_limit", tags)
 
 			if tc.statExpected {
-				require.EqualValues(t, m1.LastValue(), j.schemaHandle.CurrentColumnsCount(tableName))
+				require.EqualValues(t, m1.LastValue(), len(j.schemaHandle.schemaInWarehouse[tableName]))
 				require.EqualValues(t, m2.LastValue(), tc.columnCountLimit)
 			} else {
 				require.Nil(t, m1)
